refactor(cli): inject global analytics props on the shared sendEvent path
Previously getGlobalAnalyticsProps() (cli_version, os_platform, os_arch,
os_release, etc.) was only spread by trackEvent(). The ~40 direct
sendEvent() callers (login, org, app, bundle, channel, build, onboarding)
forwarded events to /private/events without any of these, so OS/version
segmentation was incomplete.

Move the props + invocation-source state into a standalone
analytics/global-props module (avoids a utils <-> track circular import),
inject them inside sendEvent() — the single choke point all events pass
through — and drop the now-redundant spread in trackEvent(). Caller tags
win on key conflicts. track.ts re-exports the symbols for existing import
sites.

Tests assert the globals now ride on the direct markSnag path and strip
them from the exact caller-tag matches.
